Place slice in its own module
